rename package to hadoop
