+++ clients/filesystem-hadoop3/src/test/java/com/datastrato/gravitino/filesystem/hadoop/MockServerTestBase.java
@@ -2,7 +2,7 @@
  * Copyright 2024 Datastrato Pvt Ltd.
  * This software is licensed under the Apache License version 2.
  */
-package com.datastrato.gravitino.filesystem.hadoop3;
+package com.datastrato.gravitino.filesystem.hadoop;
 
 import static org.apache.hc.core5.http.HttpStatus.SC_OK;
